#email_spammer: process single item
