+++ src/api/v1/createEmailsAction.js
@@ -16,36 +16,6 @@ const createItem = ({to, subject, html}) => {
     return item;
 };
 
-const processItem = (item) => {
-    item.status = "pending";
-
-    const transporter = nodemailer.createTransport({
-        service: 'gmail',
-        auth: {
-            user: 'mrriddick7@gmail.com',
-            pass: process.env.EMAIL_PASS,
-        }
-    });
-
-    const mailOptions = {
-        from: 'mrriddick7@gmail.com',
-        to: to,
-        subject: 'Subject of your email 2',
-        html: '<p>Your html here</p>',
-        text: 'Your text',
-    };
-
-    transporter.sendMail(mailOptions, function (err, info) {
-        if(err) {
-            item.status = "error";
-            console.log("err", err);
-        } else {
-            item.status = "success";
-            console.log("info", info);
-        }
-    });
-
-};
 
 module.exports = async (ctx) => {
     console.log(ctx.request.body);
